Improve user experience when forminator is called to destroy
If sap_state == 'absent', then forminator calls `terraform destroy`. This
changes the message to "apply terraform" instead of "deploy infrastructure
with terraform", since that is more accurate.

This also aborts the play in such cases, otherwise the playbook fails with
an error since its next tasks are to build an inventory and wait for hosts
that no longer exist.