Remove variables from playbooks
Variables defined as `foo: '{{ foo }}'` can cause problems when they are
defined with lower precendence, such as inventory variables. This can cause
a recursive loop as Ansible tries to find the value. This is being removed
to allow inventory variables to work in AWX, and they can also be defined
in vars.