Improve "development mode" for AWX image
This adds the ability to pass a service account email to the image build machine
so that AWX can be tested with a service account. It also adds a task at the end
of postinstall.yml to fail intentionally to prevent the image from being created.
When this is combined with the Packer arguments `-on-error=abort` and
`-var keep-inventory-file=true`, then Packer will exit without deleting the machine
and Ansible can later be rerun against the same machine, and the AWX UI can be
checked for correctness.