Add a network tag variable to all stacks
Because AWX doesn't have a widget for entering a list of items, there will be
one network_tag variable in all stacks, and variables such as
sap_hana_network_tags will be derived from it. If not given, it defaults to
"sap-allow-all". The SAP-Prerequisites stack has been modified to create this
tag.