Fix intermittent error with sap-pre-reqs Terraform module
In some cases, this could fail with an error when var.subnetwork_project was
passed as an empty string, as AWX does unless the user fills in the field.
This change ensures that the google_project_iam_member resources only attempt
to be created when var.subnetwork_project and var.project_id differ and
var.subnetwork_project is not an empty string.