Add assertions for high level roles
A new task file, assertions.yml, has been added to all of the high level roles,
and stack playbooks now do an `include_role` with `tasks_from: assertions` as
the last play.

This also modifies the outputs for HANA scaleout to include an inventory group
with all machines, and the names of all worker machines. This enables the
assertions to run on all machines at the same time, and to use the worker
machine names to match with the output of the `sapcontrol` command.