Add ensa_version variable to sap_product_vars
The assertions need to differ when a product uses ENSA1 vs ENSA2. For example,
with ENSA1, the output of GetProcessList will be something like:

enserver, EnqueueServer, GREEN, Running

and with ENSA2 it is:

enq_server, Enqueue Server 2, GREEN, Running

The value of ensa_version is also now passed into pacemaker-netweaver to set
sap_ensa_version which previously assumed a default of '2'.