Optionally add AAS instances when building NetWeaver-HA
This adds a new variable `sap_nw_as_num_instances` which determines the number
of application server instances. The number defaults to `1` if not provided.
The first instance is the PAS and additional instances are AAS.

The nw-ha Terraform module was modified to enable this change, and some
cleanups were done in the process.