Change Terraform inventory outputs
This changes Terraform inventory outputs from a map such as
{ hana = ["10.1.2.3", "10.1.2.4"] } to an array of maps, each map representing
one host in the inventory. The new format is:

[
  {
    host = "10.1.2.3",
    groups = ["hana"],
    vars = {
      is_primary = true,
    }
  }
]

The addition of vars allows simplifying the Ansible and makes it easier to use
Ansible without Terraform. For example, the var `is_primary` above can be defined
in the Terraform output or simply as a host var.

This change removes reliance on host naming schemes to determine the type of
instance, for example instead of relying on a naming scheme to determine for HANA
scaleout worker nodes, an inventory variable `sap_hana_is_worker` is used instead.

This also removes self links as hostnames within Ansible, instead filtering
hostvars based on the presence of inventory variables. For example, instead of
defining `sap_hana_worker_node_names` from Terraform outputs as an array of
self links, the Ansible hostvars are filtered to get `ansible_hostname` for hosts
that have `sap_hana_is_worker` defined with value `true`. This works when
defining `sap_hana_is_worker` from a Terraform inventory output or within an
inventory without Terraform.