Remove usage of storage-bkup role
The `storage-bkup` role has been replaced with `storage` (on which `storage-bkup`
is based), `nfs-client`, and `nfs-server`.

The `nfs-client` role can now mount NFS with either fstab or autofs. The
roles will use autofs in preparation for using Pacemaker to handle ASCS and
ERS mounts, per the documentation at
https://cloud.google.com/solutions/sap/docs/netweaver-ha-config-sles.

The `storage-bkup` role is not yet being removed as the legacy role
`netweaver-aas-ha` is still using it and will later be updated to the new style.