Make HA and Distributed NFS mounts configurable with variables
Previously, NFS mounts were defined based on the value of `sap_nw_nfs_src`, but
now mounts for SCS, PAS, and AAS can be individually overridden by defining a
list of dictionaries with the configuration for each mount point.

The nfs-client role has been extended to handle permissions for the mounts
based on the dictionary keys `owner`, `group`, and `mode`.