Add linting capability
This commit introduces linting capabilities for this repo. Linting
is done via a container that defined in build/Dockerfile. Linting
is implemented using the following techonologies:
- ansilbe code linting via ansible-lint
- terraform code linting via terraform fmt and terraform validate
- python code linting via flake8
- shell code linting via shellcheck
In addition to the above we also check if the files have proper
license header requires by Google PSO.
This commit also introduces cloud-build configuration that is
initial step towards CI/CD for this repo.
Linting is NOT performed on the third_party/ and venv/ directories

The code has been inspired by CFT CI/CD code. Please see
https://github.com/GoogleCloudPlatform/cloud-foundation-toolkit/tree/master/infra/build/developer-tools
for more information